bt service: rework bt service
